+++ public/locales/en/translation.json
@@ -1,3 +1,57 @@
 {
-  "test": "test"
+    "Contacts": "Contacts",
+    "Required field": "Required field",
+    "aboutUs": "About us",
+    "aboutUsDeliveryInfo1stPart": "Importing products directly from China is not as convenient as cooperating with a supplier already on the territory of Ukraine. And it doesn't even matter what you buy - flashlights or wireless headphones - in bulk it is always better to cooperate with a company that you can check. In addition, there are a number of other advantages, including:",
+    "aboutUsDeliveryInfo2ndPart": "Fast delivery - no supplier located abroad can provide delivery as fast as a supplier from Ukraine. And as faster the delivery, the more confidence in the safety of the order. This is especially true when, for example, you buy wholesale glass teapots.",
+    "aboutUsDeliveryInfo3rdPart": "Good reputation - every entrepreneur knows that reputation is earned over years and can be lost very quickly. And we are not talking about reviews on the Internet, which, in fact, can always be bought - we are talking about the human factor and word of mouth radio.",
+    "aboutUsDeliveryInfo4thPart": "We try to include only the most popular offers in our product catalog, which guarantees excellent sales. Moreover, even among products of the same type, there are more successful models and less. For example, portable speakers are purchased in bulk by a lot of retailers, however , there are certain best-selling models.Affordable prices - we not only take care of choosing the most relevant models of various product groups, but also make sure that our prices are as low as possible compared to competitors.",
+    "basket": "Shopping cart",
+    "basketAddBtn": "Add to cart",
+    "basketConfirmTheOrder": "Confirm the order",
+    "basketNumberOfProducts": "Number of products",
+    "basketTakeOutBtn": "Remove from cart",
+    "basketTotal": "Total",
+    "continueShopping": "Continue shopping",
+    "deliveryMethod": "Delivery method",
+    "error": "Error",
+    "errorPageComeBackToMain": "Back to Main",
+    "errorPageError": "An error occurred",
+    "footerSlogan": "Large selection of products in various categories, from air mattresses to tablet computers",
+    "inputSearchPlaceholder": "Search product",
+    "loadingProcessLoading": "Loading...",
+    "loadingProcessNoProduct": "No products",
+    "loadingProcessProductExist": "In stock",
+    "loadingProcessProductNotExist": "Product does not exist",
+    "loadingProcessUnexpectedError": "Unexpected error",
+    "newProductLabel": "new",
+    "orderForm": "Checkout",
+    "orderFormLabelAgreement": "I consent to the processing of my personal data",
+    "orderFormLabelCity": "City",
+    "orderFormLabelName": "First name, Last name",
+    "orderFormLabelPost": "Post Office",
+    "orderFormPhoneNumber": "Phone number",
+    "orderFormPlaceholderCity": "Budapest",
+    "orderFormPlaceholderName": "Alexander Ivanov",
+    "orderFormPlaceholderPhone": "(123) 456 78 90",
+    "orderFormPlaceholderPost": "№21 Oleksandrivsky pr...",
+    "orderFormTypeOfDelivery": "Delivery Type",
+    "productDescriptionTitle": "Description",
+    "productsCatalog": "Products catalog",
+    "productsNewProducts": "New arrivals",
+    "productsTopProducts": "Top items",
+    "requiredField": "requiredField",
+    "returnAndExchangeOfProducts": "Return and exchange of products",
+    "returnInfoDetails": "Return is possible within 14 days after receipt (for products of good quality)",
+    "returnInfoObligations": "The returned item must be in good condition. That is: - it must be packed in its original whole package (not torn, not wrinkled, not cut, etc.) - the product must not be beaten, scratched or without any other defects. The product is taken to our service for repair. If this product cannot be repaired, then it is exchanged for the same or similar one (by agreement with the manager). Or for it, its cost is returned to the bank card (services of a transport company are not included in this price). Products cheaper than 100 UAH are NOT subject to exchange and return!",
+    "returnInfoShipping": "Return delivery of products is carried out by agreement.",
+    "scheduleMondayFriday": "Mon-Fri: 8.00 - 20.00",
+    "scheduleSaturday": "Saturday: 8.00 - 16.00",
+    "scheduleTitle": "Schedule",
+    "scheduleWeekend": "Sunday: Closed",
+    "shippingAndPayment": "Shipping and payment",
+    "test": "test",
+    "view": "View",
+    "viewProducts": "View products",
+    "requiredField": "Reqiered field"
 }
