+++ public/locales/ru/translation.json
@@ -1,3 +1,57 @@
 {
-  "test": "Большой выбор товаров самых различных категорий, отнадувных матрасов до планшетных компьютеров"
+    "Contacts": "Контакты",
+    "Required field": "Required field",
+    "aboutUs": "О нас",
+    "aboutUsDeliveryInfo1stPart": "Завозить товары напрямую из Китая совсем не так удобно, как сотрудничать с поставщиком уже на территории Украины. И здесь даже не важно, что вы закупаете – фонарики или беспроводные наушники – оптом всегда лучше сотрудничать с той компанией, которую можно проверить своими силами. Кроме того, есть ряд и других преимуществ, в числе которых:",
+    "aboutUsDeliveryInfo2ndPart": "Быстрая доставка – ни один поставщик находящиеся за границей, не сможет обеспечить доставку столь быструю, как поставщик из Украины. А чем быстрее доставка, тем больше уверенности в сохранности заказа. Особенно, это актуально, когда вы, например, покупаете стеклянные чайники оптом.",
+    "aboutUsDeliveryInfo3rdPart": "Хорошая репутация – каждый предприниматель знает, что репутация зарабатывается годами, а потерять ее можно очень быстро. И речь здесь не идёт об отзывах в интернете, которые, на самом деле, всегда можно купить – речь о человеческом факторе и сарафанном радио.",
+    "aboutUsDeliveryInfo4thPart": "Мы стараемся включать в каталог наших товаров лишь наиболее популярные предложения, что гарантирует отличные продажи. Более того, даже среди товаров одного типа, есть более успешные модели и менее. Например, портативные колонки оптом закупаются очень многими розничными продавцами, однако, существуют определенные, наиболее продаваемые модели. Доступные цены – мы не только заботимся о выборе наиболее актуальных моделей различных групп товаров, но и следим за тем, чтобы наши цены по возможности были ниже чем у конкурентов.",
+    "basket": "Корзина",
+    "basketAddBtn": "Добавить в корзину",
+    "basketConfirmTheOrder": "Подтвердить заказ",
+    "basketNumberOfProducts": "Количество товаров",
+    "basketTakeOutBtn": "Убрать из корзины",
+    "basketTotal": "Итого",
+    "continueShopping": "Продолжить покупки",
+    "deliveryMethod": "Способ доставки",
+    "error": "Ошибка",
+    "errorPageComeBackToMain": "Вернуться на главную",
+    "errorPageError": "Произошла ошибка",
+    "footerSlogan": "Большой выбор товаров самых различных категорий, отнадувных матрасов до планшетных компьютеров",
+    "inputSearchPlaceholder": "Поиск товара",
+    "loadingProcessLoading": "Загрузка...",
+    "loadingProcessNoProduct": "Нет товаров",
+    "loadingProcessProductExist": "В наличии",
+    "loadingProcessProductNotExist": "Продукт не существует",
+    "loadingProcessUnexpectedError": "Неизвестная ошибка",
+    "newProductLabel": "новый",
+    "orderForm": "Оформление заказа",
+    "orderFormLabelAgreement": "Я согласен (согласна) на обработку моих персональных данных",
+    "orderFormLabelCity": "Город",
+    "orderFormLabelName": "Имя, Фамилия",
+    "orderFormLabelPost": "Отделение Новой почты",
+    "orderFormPhoneNumber": "Номер телефона",
+    "orderFormPlaceholderCity": "Будапешт",
+    "orderFormPlaceholderName": "Александр Иванов",
+    "orderFormPlaceholderPhone": "(123) 456 78 90",
+    "orderFormPlaceholderPost": "№21 Олександрівський пр...",
+    "orderFormTypeOfDelivery": "Тип доставки",
+    "productDescriptionTitle": "Описание",
+    "productsCatalog": "Каталог товаров",
+    "productsNewProducts": "Новые поступления",
+    "productsTopProducts": "Топовые позиции",
+    "requiredField": "requiredField",
+    "returnAndExchangeOfProducts": "Возврат и обмен товара",
+    "returnInfoDetails": "Возврат возможен в течение 14 дней после получения (для товаров надлежащего качества)",
+    "returnInfoObligations": "Возвращаемый товар должен быть в надлежащем виде. То есть: - должен быть упакован в его родную целую упаковку (не порванную, не помятую, не разрезанную и т.п.) - товар должен быть не побит, поцарапан или без каких либо других дефектов. Товар забирается на ремонт в наш сервис. Если данный товар починить невозможно,то он обменивается на такой же или аналогичный(по договоренности с менеджером). Или за него возвращаются его стоимость на карту банка (услуги транспортной компании в эту стоимость не входит). Товар дешевле 100 грн обмену и возврату НЕ ПОДЛЕЖИТ!",
+    "returnInfoShipping": " Обратная доставка товаров осуществляется по договоренности.",
+    "scheduleMondayFriday": "Пн-Пт: 8.00 - 20.00",
+    "scheduleSaturday": "Сб: 8.00 - 16.00",
+    "scheduleTitle": "График работы",
+    "scheduleWeekend": "Вс: Выходной",
+    "shippingAndPayment": "Доставка и оплата",
+    "test": "Большой выбор товаров самых различных категорий, отнадувных матрасов до планшетных компьютеров",
+    "view": "Смотреть",
+    "viewProducts": "Смотреть товары",
+    "requiredField": "Обязательное поле"
 }
